requirements.txt
flask==3.0.0
gunicorn==21.2.0
anthropic==0.40.0
voyageai==0.3.1
chromadb==0.5.23
python-dotenv==1.0.0
requests==2.32.0
